searcher/requirements.txt
# FastAPI framework
fastapi

# ASGI server for FastAPI
uvicorn

# For RabbitMQ integration
pika
protobuf==3.20.0

# OpenTelemetry for tracing and metrics
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
opentelemetry-exporter-zipkin

# For MySQL integration with SQLAlchemy and SQLModel
python-dotenv
sqlalchemy
sqlmodel
aiomysql

